+++ sdk/typespec/typespec_client_core/CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### Breaking Changes
 
+- Removed `CustomHeaders` policy.
+
 ### Bugs Fixed
 
 ### Other Changes